requirements.txt
selenium==4.11.2
pillow==9.5.0